chore(package): move packages to root
