+++ mid/machineid_windows.go
