+++ path_tools/find_file_path_test.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 	"testing"
 
-	pathtools "github.com/guionardo/go/pkg/path_tools"
+	pathtools "github.com/guionardo/go/path_tools"
 	"github.com/stretchr/testify/require"
 )
