+++ set/scanner_valuer.go
