+++ set/set_test.go
@@ -3,7 +3,7 @@ package set_test
 import (
 	"testing"
 
-	"github.com/guionardo/go/pkg/set"
+	"github.com/guionardo/go/set"
 	"github.com/stretchr/testify/assert"
 )
